Extracted the container building to a separate class
